requirements.txt
astroid==2.15.0
autopep8==2.0.0
bidict==0.22.1
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
dill==0.3.6
dnspython==2.2.1
Flask==2.2.2
Flask-Login==0.6.2
Flask-SocketIO==5.3.2
google-auth==2.15.0
google-auth-oauthlib==0.7.1
idna==3.4
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
JSON-Datetime==0.0.6
lazy-object-proxy==1.9.0
MarkupSafe==2.1.1
mccabe==0.7.0
numpy==1.24.0
oauthlib==3.2.2
paho-mqtt==1.6.1
pandas==1.5.2
platformdirs==3.1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
PyJWT==2.6.0
pymongo==4.3.3
pypng==0.20220715.0
PyQRCode==1.2.1
python-dateutil==2.8.2
python-engineio==4.3.4
python-socketio==5.7.2
pytz==2022.7
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
simplejson==3.18.0
six==1.16.0
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.5.0
urllib3==1.26.13
Werkzeug==2.2.2
wrapt==1.15.0
